requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
beautifulsoup4==4.13.4
feedparser==6.0.11
frozenlist==1.6.0
idna==3.10
multidict==6.4.3
propcache==0.3.1
sgmllib3k==1.0.0
soupsieve==2.7
typing_extensions==4.13.2
yarl==1.20.0
discord.py
requests
python-dotenv
